feat(move onevent): move onEvent from CourierProvider to Inbox so we can access hooks better
having the function on CourierProvider made it hard to use things like useInbox
